requirements.txt
python-dotenv
black
flake8
py-cord
